chore(qrm): refactor cpu/memory read-only state
